Support FLAC files, stop pulling musicbrainz data
Assume that all of the files were tagged properly, so just grab the
metadata out of them instead of querying MusicBrainz constantly.

MP3 is now a free file format, so there's no need to transcode to ogg
anymore really. Instead, we can transcode our FLAC files to MP3 as
needed.